Ajout de la sous commande info
